3-DevOps/sql-python-portfolio/requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
psycopg2-binary
Werkzeug
Faker
